requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
boto3==1.23.9
botocore==1.26.9
certifi==2022.5.18.1
configparser==5.2.0
Django==3.2
django-admin==2.0.1
django-bootstrap-v5==1.0.11
django-debug-toolbar==3.4.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-minio-storage==0.3.10
django-phonenumber-field==6.1.0
django-six==1.0.4
django-storages==1.12.3
jmespath==1.0.0
minio==6.0.2
phonenumbers==8.12.48
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
screen==1.0.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
xlwt==1.3.0
